feat: purge underViews option

+++ packages/api/src/modules/storage/storage.service.ts
@@ -1,9 +1,9 @@
-import { GetObjectCommand, PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
+import { DeleteObjectCommand, GetObjectCommand, PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
 import type { EntityRepository } from "@mikro-orm/postgresql";
 import { CreateRequestContext, EntityManager, type FilterQuery } from "@mikro-orm/core";
 import { InjectRepository } from "@mikro-orm/nestjs";
-import { Injectable, Logger } from "@nestjs/common";
-import { Cron, CronExpression, Interval } from "@nestjs/schedule";
+import { Injectable, Logger, type OnApplicationBootstrap } from "@nestjs/common";
+import { Interval } from "@nestjs/schedule";
 import crypto from "crypto";
 import { once } from "events";
 import fs, { existsSync, mkdirSync, rmdirSync } from "fs";
@@ -20,9 +20,11 @@ import { setTimeout } from "timers/promises";
 import ms from "ms";
 import { dedupe } from "../../helpers/dedupe.js";
 import { ThumbnailService } from "../thumbnail/thumbnail.service.js";
+import { DateTime } from "luxon";
+import bytes from "bytes";
 
 @Injectable()
-export class StorageService {
+export class StorageService implements OnApplicationBootstrap {
   @InjectRepository("FileEntity") private fileRepo: EntityRepository<FileEntity>;
 
   private readonly createdPaths = new Set();
@@ -88,16 +90,29 @@ export class StorageService {
     }
   }
 
-  async delete(hash: string) {
-    try {
-      this.logger.debug(`Deleting "${hash}" from disk`);
-      const filePath = this.getPathFromHash(hash);
-      await fs.promises.unlink(filePath);
-    } catch (error: any) {
-      if (error.code === "ENOENT") {
-        return;
+  async delete(file: { hash: string; external: boolean }) {
+    this.logger.debug(`Deleting file "${file.hash}"`);
+    if (file.external) {
+      if (!this.s3Client || !config.externalStorage) {
+        throw new Error("File is external but external storage is not configured");
       }
 
+      const s3Key = this.getFolderFromHash(file.hash);
+      await this.s3Client.send(
+        new DeleteObjectCommand({
+          Bucket: config.externalStorage.bucket,
+          Key: s3Key,
+        }),
+      );
+
+      return;
+    }
+
+    try {
+      const filePath = this.getPathFromHash(file.hash);
+      await fs.promises.unlink(filePath);
+    } catch (error: any) {
+      if (error.code === "ENOENT") return;
       throw error;
     }
   }
@@ -150,7 +165,7 @@ export class StorageService {
   @Interval(ms("5s"))
   @CreateRequestContext()
   @dedupe()
-  async uploadToExternalStorage(): Promise<void> {
+  protected async uploadToExternalStorage(): Promise<void> {
     if (!config.externalStorage || !this.s3Client) return;
 
     const filter: FilterQuery<FileEntity>[] = [];
@@ -165,18 +180,23 @@ export class StorageService {
       filter.push({ createdAt: { $lte: decayDate } });
     }
 
-    const file = await this.fileRepo.findOne({
-      external: false,
-      externalError: null,
-      $and: filter,
-      // we want to generate thumbnails while the file is not external, because it
-      // saves s3 calls which may cost money.
-      $or: [
-        { thumbnail: { $ne: null } },
-        { thumbnailError: { $ne: null } },
-        { type: { $nin: [...ThumbnailService.IMAGE_TYPES, ...ThumbnailService.VIDEO_TYPES] } },
-      ],
-    });
+    const file = await this.fileRepo.findOne(
+      {
+        external: false,
+        externalError: null,
+        $and: filter,
+        // we want to generate thumbnails while the file is not external, because it
+        // saves s3 calls which may cost money.
+        $or: [
+          { thumbnail: { $ne: null } },
+          { thumbnailError: { $ne: null } },
+          { type: { $nin: [...ThumbnailService.IMAGE_TYPES, ...ThumbnailService.VIDEO_TYPES] } },
+        ],
+      },
+      {
+        fields: ["type", "size", "hash", "external", "externalError"],
+      },
+    );
 
     if (!file) {
       await setTimeout(ms("5m"));
@@ -210,6 +230,58 @@ export class StorageService {
       this.logger.error("This is a bug, report it on github!");
       this.logger.warn("Retrying external store uploads in 30 minutes");
       await setTimeout(ms("30m"));
+    }
+  }
+
+  @Interval(ms("5s"))
+  @CreateRequestContext()
+  @dedupe()
+  protected async purgeFiles() {
+    if (!config.purge) return;
+    const createdBefore = new Date(Date.now() - config.purge.afterTime);
+    const filters: FilterQuery<FileEntity> = [];
+    if (config.purge.underViews) {
+      filters.push({
+        views: { $lt: config.purge.underViews },
+      });
+    }
+
+    const files = await this.fileRepo.find(
+      {
+        $and: filters,
+        size: {
+          $gte: config.purge.overLimit,
+        },
+        createdAt: {
+          $lte: createdBefore,
+        },
+      },
+      {
+        limit: 10,
+        fields: ["size", "createdAt", "external", "hash"],
+      },
+    );
+
+    for (const file of files) {
+      const size = bytes.format(file.size);
+      const age = DateTime.fromJSDate(file.createdAt).toRelative();
+      await this.em.removeAndFlush(file);
+      await this.delete(file);
+      this.logger.log(`Purging ${file.id} (${size}, ${age})`);
+    }
+
+    if (files[0]) {
+      this.logger.log(`Purged ${files.length} files`);
+    } else {
+      await setTimeout(ms("5m"));
+    }
+  }
+
+  onApplicationBootstrap() {
+    if (config.purge) {
+      const size = bytes.format(config.purge.overLimit);
+      const age = DateTime.local().minus(config.purge.afterTime).toRelative();
+      this.logger.warn(`Purging files is enabled for files over ${size} uploaded more than ${age}.`);
     }
   }
 }
